+++ src/banner/FPSWebPartClass/BuildBannerPropsX2.ts
@@ -187,13 +187,13 @@ export function mainWebPartRenderBannerSetupX( setup: IMainWPBannerSetupX ) : IW
 
     const MinFPProps: any = setup.main.properties;
 
-    const saveCommands: any = _FieldPanelDesignMode === 'Disabled' ? null : setup.main._saveFieldPanelCommandsFunction;
-    const saveViews: any = _FieldPanelDesignMode === 'Disabled' ? null : setup.main._saveFieldPanelViewsFunction;
+    const saveCommands: any = _FieldPanelDesignMode === 'Disabled' ? null : setup.main._saveFieldPanelCommandsFunction.bind( setup.main );
+    const saveViews: any = _FieldPanelDesignMode === 'Disabled' ? null : setup.main._saveFieldPanelViewsFunction.bind( setup.main );
 
     const fieldPanelProps: IMinWPFieldPanelProps = {
       displayMode: displayMode,
       lists: [{
-        webUrl: _FieldPanelWebProp ? MinFPProps[ _FieldPanelWebProp ] : '' ,
+        webUrl: _FieldPanelWebProp && MinFPProps[ _FieldPanelWebProp ] ? MinFPProps[ _FieldPanelWebProp ] : pageContext.web.absoluteUrl ,
         listTitle: _FieldPanelListProp ? MinFPProps[ _FieldPanelListProp ] : 'Documents' ,
       }],
       designMode: _FieldPanelDesignMode,
@@ -205,6 +205,17 @@ export function mainWebPartRenderBannerSetupX( setup: IMainWPBannerSetupX ) : IW
 
     const startTime = new Date();
     const refreshId = startTime.toISOString().replace('T', ' T'); // + ' ~ ' + startTime.toLocaleTimeString();
+
+    /***
+ *    d8888b.  .d8b.  d8b   db d8b   db d88888b d8888b.      d8888b. d8888b.  .d88b.  d8888b. .d8888. 
+ *    88  `8D d8' `8b 888o  88 888o  88 88'     88  `8D      88  `8D 88  `8D .8P  Y8. 88  `8D 88'  YP 
+ *    88oooY' 88ooo88 88V8o 88 88V8o 88 88ooooo 88oobY'      88oodD' 88oobY' 88    88 88oodD' `8bo.   
+ *    88~~~b. 88~~~88 88 V8o88 88 V8o88 88~~~~~ 88`8b        88~~~   88`8b   88    88 88~~~     `Y8b. 
+ *    88   8D 88   88 88  V888 88  V888 88.     88 `88.      88      88 `88. `8b  d8' 88      db   8D 
+ *    Y8888P' YP   YP VP   V8P VP   V8P Y88888P 88   YD      88      88   YD  `Y88P'  88      `8888Y' 
+ *                                                                                                    
+ *                                                                                                    
+ */
 
     const bannerProps: IWebpartBannerProps = {